requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-speech==2.22.0
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
openai==1.3.5
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
rsa==4.9
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0

nltk~=3.8.1
pydub~=0.25.1